requirements.txt
asgiref==3.5.2
Django==4.0.6
django-ckeditor==6.4.2
django-environ==0.9.0
django-filter==22.1
django-jazzmin==2.5.0
django-js-asset==2.0.0
django-modeltranslation==0.18.4
django-tagging==0.5.0
djangorestframework==3.13.1
env==0.1.0
Markdown==3.4.1
Pillow==9.2.0
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
